Update Dockerfile
Hi, this is Mizzlr. I have restructured your Docker file to reduce the image size and number of layers. It's a good practice to run multiple shell commands in one RUN layer as it reduces the number of layers and helps reduce the image size by removing apt lists at the end. The original image size was 520.9MB, but with the modified Docker file, it's just 286 MB. That is about 50 % reduction in size. I have docker-built and tested the above Docker file. Please accept my proposal for change file.